# routes/user.js
const router = require("express").Router()
const {verifyToken,verifyTokenAdmin, verifyTokenAndAuth} = require("../middleware/VerifyToken")
const CryptoJs = require("crypto-js")
const User = require("../models/User")
//UPDATE USER
router.put("/:id", verifyTokenAndAuth, async (req,res) => {

    if(req.body.password ){ // IF EXISTS ENCRYPT GIVEN PASSWORD

        req.body.password = CryptoJs.AES.encrypt(
            req.body.password,
            process.env.JWT_SEC
        ).toString()
    }

    try {
        const updatedUser = await User.findByIdAndUpdate(req.params.id, {
            $set: req.body
        }, {new: true})
        res.status(200).send(updatedUser)
    } catch (e) {
        res.status(500).json(e)
    }

})

//DELETE
router.delete("/:id", verifyTokenAndAuth, async (req, res) => {
    try {
        await User.findByIdAndDelete(req.params.id);
        res.status(200).json("User has been deleted...");
    } catch (err) {
        res.status(500).json(err);
    }
});

//GET USER
router.get("/find/:id", verifyTokenAdmin, async (req, res) => {
    try {
        const user = await User.findById(req.params.id);
        const { password, ...others } = user._doc;
        res.status(200).json(others);
    } catch (err) {
        res.status(500).json(err);
    }
});


//GET ALL USER
router.get("/", verifyTokenAdmin , async (req, res) => {
    const query = req.query.new;
    try {
        const users = query
            ? await User.find().sort({ _id: -1 } ).limit(5)
            : await User.find();
        res.status(200).json(users);
    } catch (err) {
        res.status(500).json(err);
    }
});

//GET USER STATS

router.get("/stats", verifyTokenAdmin, async (req, res) => {

    const date = new Date();
    const lastYear = new Date(date.setFullYear(date.getFullYear() - 1));

    try {
        //aggregate --> groups models
        const data = await User.aggregate([
            { $match: { createdAt: { $gte: lastYear } } },
            {
                $project: {
                    month: { $month: "$createdAt" },
                },
            },
            {
                $group: {
                    _id: "$month",
                    total: { $sum: 1 },
                },
            },
        ]);
        res.status(200).json(data)
    } catch (err) {
        res.status(500).json(err);
    }
});

module.exports = router